requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.40
requests>=2.31.0
python-dotenv>=1.0.0
cachetools>=5.3.0
apscheduler>=3.10.0
scikit-learn>=1.4.0
hypothesis>=6.100.0
pytest>=8.0.0
pytest-mock>=3.14.0
httpx>=0.27.0
